newt: improve gc track in partial replication
currently, gc tracks the commits belonging to each shard; with partial
replication, there are commands committed that "belong" (i.e. were
initially targetted) to other shards. this changes makes it that we
don't track the commit of such commands and remove gc them right on
commit. once we implement recovery in partial replication, this
mechanism should be revisited.