fanotch_run: reduce the number of connections per process
previously we were connecting each process to every other process, even
though not all possible pair of connections are used. now, we only
connect each process to every process in its region (so that it has a
connection to every other shard) and to every process in its shard (that
leave in the other regions).