fantoch: improve Executor trait to reduce allocatinos
Similarly to Protocol, Executor trait used to require allocating a
vector to store executor results. Now, implementations of Executor keep
one such vector to which results are pushed to. A new method
'to_clients' extracts a single executor result (if one exists).
Similarly to the changes in the Protocol trait, a convenience method
'to_clients_iter' is added.