atlas: remove "reducing dependencies in slow path" optimization
this optimization is not compatible with compact representations of
dependencies (when enabled, it's possible that two conflicing commands are
committed without one being the a dependency of the other), such as the one
we're using now (direct dependencies) introduced in #137.
it is also not compatible with the one we were using previouly (map with the
highest conflicting command per replica) (well, it might be compatible, but we
would need to fix the computation of the threshold union in 'threshold-rs' to be
aware of the be aware of the issue, as currently only worries about aggrement).
for the above reasons, we disable the optimization.