caesar: fix nasty bug
We were upgrading a read guard to a write guard by dropping the read
guard and acquiring the write guard. (Protip: don't do this)
This would be fine if this this upgrade was atomic, which is not. So, it
is possible that a thread does something in between these two actions,
and the code was assuming that that would not happen. Solution: use a
Mutex.